requirements.txt
langchain-anthropic
langgraph
yfinance>=1.2.0
pydantic
python-dotenv
flask
typing-extensions
